Replaced IntOpenHashMap with IntOpenHashTable
