+++ core/src/main/java/hivemall/fm/FMIntFeatureMapModel.java
@@ -19,7 +19,7 @@
 package hivemall.fm;
 
 import hivemall.utils.collections.maps.Int2FloatOpenHashTable;
-import hivemall.utils.collections.maps.IntOpenHashMap;
+import hivemall.utils.collections.maps.IntOpenHashTable;
 
 import java.util.Arrays;
 
@@ -33,7 +33,7 @@ public final class FMIntFeatureMapModel extends FactorizationMachineModel {
     // LEARNING PARAMS
     private float _w0;
     private final Int2FloatOpenHashTable _w;
-    private final IntOpenHashMap<float[]> _V;
+    private final IntOpenHashTable<float[]> _V;
 
     private int _minIndex, _maxIndex;
 
@@ -42,7 +42,7 @@ public final class FMIntFeatureMapModel extends FactorizationMachineModel {
         this._w0 = 0.f;
         this._w = new Int2FloatOpenHashTable(DEFAULT_MAPSIZE);
         _w.defaultReturnValue(0.f);
-        this._V = new IntOpenHashMap<float[]>(DEFAULT_MAPSIZE);
+        this._V = new IntOpenHashTable<float[]>(DEFAULT_MAPSIZE);
         this._minIndex = 0;
         this._maxIndex = 0;
     }
