+++ core/src/main/java/hivemall/utils/collections/maps/IntOpenHashTable.java
@@ -25,54 +25,63 @@ import java.io.IOException;
 import java.io.ObjectInput;
 import java.io.ObjectOutput;
 import java.util.Arrays;
-import java.util.HashMap;
 
 import javax.annotation.Nonnull;
 
 /**
- * An open-addressing hash table with double-hashing that requires less memory to {@link HashMap}.
+ * An open-addressing hash table with double hashing
+ *
+ * @see http://en.wikipedia.org/wiki/Double_hashing
  */
 public final class IntOpenHashTable<V> implements Externalizable {
+    private static final long serialVersionUID = -8162355845665353513L;
 
     public static final float DEFAULT_LOAD_FACTOR = 0.75f;
     public static final float DEFAULT_GROW_FACTOR = 2.0f;
 
-    public static final byte FREE = 0;
-    public static final byte FULL = 1;
-    public static final byte REMOVED = 2;
+    protected static final byte FREE = 0;
+    protected static final byte FULL = 1;
+    protected static final byte REMOVED = 2;
 
     protected/* final */float _loadFactor;
     protected/* final */float _growFactor;
 
-    protected int _used = 0;
+    protected int _used;
     protected int _threshold;
 
     protected int[] _keys;
     protected V[] _values;
     protected byte[] _states;
 
-    public IntOpenHashTable() {} // for Externalizable
+    public IntOpenHashTable() {} // for Externalizable    
 
     public IntOpenHashTable(int size) {
-        this(size, DEFAULT_LOAD_FACTOR, DEFAULT_GROW_FACTOR);
+        this(size, DEFAULT_LOAD_FACTOR, DEFAULT_GROW_FACTOR, true);
+    }
+
+    public IntOpenHashTable(int size, float loadFactor, float growFactor) {
+        this(size, loadFactor, growFactor, true);
     }
 
     @SuppressWarnings("unchecked")
-    public IntOpenHashTable(int size, float loadFactor, float growFactor) {
+    protected IntOpenHashTable(int size, float loadFactor, float growFactor, boolean forcePrime) {
         if (size < 1) {
             throw new IllegalArgumentException();
         }
         this._loadFactor = loadFactor;
         this._growFactor = growFactor;
-        int actualSize = Primes.findLeastPrimeNumber(size);
+        this._used = 0;
+        int actualSize = forcePrime ? Primes.findLeastPrimeNumber(size) : size;
+        this._threshold = Math.round(actualSize * _loadFactor);
         this._keys = new int[actualSize];
         this._values = (V[]) new Object[actualSize];
         this._states = new byte[actualSize];
-        this._threshold = Math.round(actualSize * _loadFactor);
     }
 
     public IntOpenHashTable(@Nonnull int[] keys, @Nonnull V[] values, @Nonnull byte[] states,
             int used) {
+        this._loadFactor = DEFAULT_LOAD_FACTOR;
+        this._growFactor = DEFAULT_GROW_FACTOR;
         this._used = used;
         this._threshold = keys.length;
         this._keys = keys;
@@ -80,14 +89,17 @@ public final class IntOpenHashTable<V> implements Externalizable {
         this._states = states;
     }
 
+    @Nonnull
     public int[] getKeys() {
         return _keys;
     }
 
+    @Nonnull
     public Object[] getValues() {
         return _values;
     }
 
+    @Nonnull
     public byte[] getStates() {
         return _states;
     }
@@ -109,7 +121,7 @@ public final class IntOpenHashTable<V> implements Externalizable {
         int keyLength = _keys.length;
         int keyIdx = hash % keyLength;
 
-        boolean expanded = preAddEntry(keyIdx);
+        final boolean expanded = preAddEntry(keyIdx);
         if (expanded) {
             keyLength = _keys.length;
             keyIdx = hash % keyLength;
@@ -119,14 +131,14 @@ public final class IntOpenHashTable<V> implements Externalizable {
         final V[] values = _values;
         final byte[] states = _states;
 
-        if (states[keyIdx] == FULL) {
+        if (states[keyIdx] == FULL) {// double hashing
             if (keys[keyIdx] == key) {
                 V old = values[keyIdx];
                 values[keyIdx] = value;
                 return old;
             }
             // try second hash
-            int decr = 1 + (hash % (keyLength - 2));
+            final int decr = 1 + (hash % (keyLength - 2));
             for (;;) {
                 keyIdx -= decr;
                 if (keyIdx < 0) {
@@ -149,10 +161,50 @@ public final class IntOpenHashTable<V> implements Externalizable {
         return null;
     }
 
+    public V putIfAbsent(final int key, final V value) {
+        final int hash = keyHash(key);
+        int keyLength = _keys.length;
+        int keyIdx = hash % keyLength;
+
+        final boolean expanded = preAddEntry(keyIdx);
+        if (expanded) {
+            keyLength = _keys.length;
+            keyIdx = hash % keyLength;
+        }
+
+        final int[] keys = _keys;
+        final V[] values = _values;
+        final byte[] states = _states;
+
+        if (states[keyIdx] == FULL) {// second hashing
+            if (keys[keyIdx] == key) {
+                return values[keyIdx];
+            }
+            // try second hash
+            final int decr = 1 + (hash % (keyLength - 2));
+            for (;;) {
+                keyIdx -= decr;
+                if (keyIdx < 0) {
+                    keyIdx += keyLength;
+                }
+                if (isFree(keyIdx, key)) {
+                    break;
+                }
+                if (states[keyIdx] == FULL && keys[keyIdx] == key) {
+                    return values[keyIdx];
+                }
+            }
+        }
+        keys[keyIdx] = key;
+        values[keyIdx] = value;
+        states[keyIdx] = FULL;
+        _used++;
+        return null;
+    }
 
     /** Return weather the required slot is free for new entry */
-    protected boolean isFree(int index, int key) {
-        byte stat = _states[index];
+    protected boolean isFree(final int index, final int key) {
+        final byte stat = _states[index];
         if (stat == FREE) {
             return true;
         }
@@ -163,8 +215,8 @@ public final class IntOpenHashTable<V> implements Externalizable {
     }
 
     /** @return expanded or not */
-    protected boolean preAddEntry(int index) {
-        if ((_used + 1) >= _threshold) {// filled enough
+    protected boolean preAddEntry(final int index) {
+        if ((_used + 1) >= _threshold) {// too filled
             int newCapacity = Math.round(_keys.length * _growFactor);
             ensureCapacity(newCapacity);
             return true;
@@ -172,7 +224,7 @@ public final class IntOpenHashTable<V> implements Externalizable {
         return false;
     }
 
-    protected int findKey(final int key) {
+    private int findKey(final int key) {
         final int[] keys = _keys;
         final byte[] states = _states;
         final int keyLength = keys.length;
@@ -184,7 +236,7 @@ public final class IntOpenHashTable<V> implements Externalizable {
                 return keyIdx;
             }
             // try second hash
-            int decr = 1 + (hash % (keyLength - 2));
+            final int decr = 1 + (hash % (keyLength - 2));
             for (;;) {
                 keyIdx -= decr;
                 if (keyIdx < 0) {
@@ -217,7 +269,7 @@ public final class IntOpenHashTable<V> implements Externalizable {
                 return old;
             }
             //  second hash
-            int decr = 1 + (hash % (keyLength - 2));
+            final int decr = 1 + (hash % (keyLength - 2));
             for (;;) {
                 keyIdx -= decr;
                 if (keyIdx < 0) {
@@ -255,28 +307,49 @@ public final class IntOpenHashTable<V> implements Externalizable {
         this._used = 0;
     }
 
-    protected void ensureCapacity(int newCapacity) {
+    @Override
+    public String toString() {
+        int len = size() * 10 + 2;
+        final StringBuilder buf = new StringBuilder(len);
+        buf.append('{');
+        final IMapIterator<V> i = entries();
+        while (i.next() != -1) {
+            buf.append(i.getKey());
+            buf.append('=');
+            buf.append(i.getValue());
+            if (i.hasNext()) {
+                buf.append(',');
+            }
+        }
+        buf.append('}');
+        return buf.toString();
+    }
+
+    private void ensureCapacity(final int newCapacity) {
         int prime = Primes.findLeastPrimeNumber(newCapacity);
         rehash(prime);
         this._threshold = Math.round(prime * _loadFactor);
     }
 
     @SuppressWarnings("unchecked")
-    private void rehash(int newCapacity) {
+    private void rehash(final int newCapacity) {
         int oldCapacity = _keys.length;
         if (newCapacity <= oldCapacity) {
             throw new IllegalArgumentException("new: " + newCapacity + ", old: " + oldCapacity);
         }
+        final int[] oldKeys = _keys;
+        final V[] oldValues = _values;
+        final byte[] oldStates = _states;
         final int[] newkeys = new int[newCapacity];
         final V[] newValues = (V[]) new Object[newCapacity];
         final byte[] newStates = new byte[newCapacity];
         int used = 0;
         for (int i = 0; i < oldCapacity; i++) {
-            if (_states[i] == FULL) {
+            if (oldStates[i] == FULL) {
                 used++;
-                int k = _keys[i];
-                V v = _values[i];
-                int hash = keyHash(k);
+                final int k = oldKeys[i];
+                final V v = oldValues[i];
+                final int hash = keyHash(k);
                 int keyIdx = hash % newCapacity;
                 if (newStates[keyIdx] == FULL) {// second hashing
                     int decr = 1 + (hash % (newCapacity - 2));
@@ -287,9 +360,9 @@ public final class IntOpenHashTable<V> implements Externalizable {
                         }
                     }
                 }
-                newStates[keyIdx] = FULL;
                 newkeys[keyIdx] = k;
                 newValues[keyIdx] = v;
+                newStates[keyIdx] = FULL;
             }
         }
         this._keys = newkeys;
@@ -303,7 +376,7 @@ public final class IntOpenHashTable<V> implements Externalizable {
     }
 
     @Override
-    public void writeExternal(ObjectOutput out) throws IOException {
+    public void writeExternal(@Nonnull final ObjectOutput out) throws IOException {
         out.writeFloat(_loadFactor);
         out.writeFloat(_growFactor);
         out.writeInt(_used);
@@ -319,8 +392,8 @@ public final class IntOpenHashTable<V> implements Externalizable {
     }
 
     @SuppressWarnings("unchecked")
-    @Override
-    public void readExternal(ObjectInput in) throws IOException, ClassNotFoundException {
+    public void readExternal(@Nonnull final ObjectInput in) throws IOException,
+            ClassNotFoundException {
         this._loadFactor = in.readFloat();
         this._growFactor = in.readFloat();
         this._used = in.readInt();
